+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR0ABMN27523HASF62R8KS04"
+        "01KR0QJ3NPQ09W0GQ5VX8MX92P"
     )
     r"""The unique identifier of the data source"""