+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR0ABMMZPT1RB9XPJCFPX507"
+        "01KR0QJ3NK3Z2V5YNQ85NGYQPY"
     )
     r"""The unique identifier of the data source"""