+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR0ABMN4XWGQW7QHJXMCZD7Y"
+        "01KR0QJ3NSAR5QS1TC3R2FBPW9"
     )
     r"""The unique identifier of the data source"""