+++ packages/orq-rc/RELEASES.md
@@ -7619,3 +7619,13 @@ Based on:
 - [python v4.9.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.29 - packages/orq-rc
+
+## 2026-05-07 08:09:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.30 - packages/orq-rc